Renamed files, added .txt files, new Ranglijst case code Renamed FileReader & FileWriter to ReadFiles & WriteFiles because the name was interfering with the variables of the same name.
Added readGameListData() to ReadFiles class that returns the game file so that it can be read.

Added some test code that uses the ReadFiles class to read from the gamesList.txt file and display the list of games. (WIP)

Cleaned up the code a bit around the initial launch code and added some lines.